requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
EbookLib==0.18
fastapi==0.110.3
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gradio==5.1.0
gradio_client==1.4.0
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
loguru==0.7.3
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.3.2
narwhals==1.33.0
nest-asyncio==1.6.0
nltk==3.9.1
npm==0.1.1
numpy==1.26.4
openai==1.69.0
optional-django==0.1.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
ruff==0.11.2
aiosqlite>=0.17
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.37.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.0.0
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.29.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==12.0
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.2
XlsxWriter==3.2.2
yarl==1.18.3

# 添加 style_validator.py 依赖
sentence-transformers
langdetect

# Add libraries for file parsing
pymupdf
ebooklib
beautifulsoup4
lxml

# Add database dependencies
sqlalchemy>=2.0

# Add SSE dependencies
sse-starlette

mcp-server-fetch